Create module folder and move class
