+++ index.html
@@ -8,17 +8,17 @@
     <link href='https://fonts.googleapis.com/css?family=Poppins' rel='stylesheet'>
     <title>Document</title>
 </head>
-
 <body>
+    
     <header>
            <nav>
                <div class="img-container">
                 <img src="book-logo.png"/> 
                </div>
             <ul>
-                <li><a href="#" class="list active" onclick="getPageContent('.list')" >List</a></li>
-                <li><a href="#" class="add-new" onclick="getPageContent('.add-new')">Add new</a></li>
-                <li><a href="#" class="contact" onclick="getPageContent('.contact')">Contact</a></li>
+                <li><a href="#" class="list active">List</a></li>
+                <li><a href="#" class="add-new">Add new</a></li>
+                <li><a href="#" class="contact">Contact</a></li>
             </ul>
         </nav>
     </header>
@@ -54,9 +54,9 @@
                 You can reach out to us!
                 </p>
                 <ul>
-                <li> Our e-mail: <a>alishahsafdari905@gmail.com</a></li>
+                <li>Our e-mail: <a>alishahsafdari905@gmail.com</a></li>
                 <li>Our phone number: +93744716508 </li>
-                <li> Our address : Microverse Students</li>
+                <li>Our address : Microverse Students</li>
                 </ul>
                 </div>
                 </div>
@@ -67,6 +67,6 @@
     </footer>
 
 </body>
-<script src="index.js"></script>
+<script src="index.js" type="module"></script>
 
 </html>
